+++ src/resty/ljsonschema/init.lua
@@ -29,6 +29,11 @@ local tonumber = tonumber
 local default_null = nil        -- default null token
 local default_array_mt = nil    -- default array_mt metatable
 local default_match_pattern     -- default reg-ex engine to use
+local default_len do
+  local utf_len = require("resty.ljsonschema.utf8").len  -- string length function
+  default_len = function(s) return utf_len(s) or #s end -- add fallback to bytecount in case of invalid utf8
+end
+
 do
   local ok, cjson = pcall(require, 'cjson')
   if ok then
@@ -799,16 +804,20 @@ generate_validator = function(ctx, schema)
 
   if schema.minLength or schema.maxLength or schema.pattern or schema.format then
     ctx:stmt(sformat('if %s == "string" then', datatype))
+    ctx:stmt(sformat('  local length = %s(%s)', ctx:libfunc('custom.str_len'), ctx:param(1)))
+    ctx:stmt(        '  if not length then') -- allows for overriding and NOT allowing invalid UTF8
+    ctx:stmt(        '    return false, "failed to get string length, invalid utf8"')
+    ctx:stmt(        '  end')
     if schema.minLength then
-      ctx:stmt(sformat('  if #%s < %d then', ctx:param(1), schema.minLength))
-      ctx:stmt(sformat('    return false, %s("string too short, expected at least %d, got %%d", #%s)',
-                       ctx:libfunc('string.format'), schema.minLength, ctx:param(1)))
+      ctx:stmt(sformat('  if length < %d then', schema.minLength))
+      ctx:stmt(sformat('    return false, %s("string too short, expected at least %d, got %%d", length)',
+                       ctx:libfunc('string.format'), schema.minLength))
       ctx:stmt(        '  end')
     end
     if schema.maxLength then
-      ctx:stmt(sformat('  if #%s > %d then', ctx:param(1), schema.maxLength))
-      ctx:stmt(sformat('    return false, %s("string too long, expected at most %d, got %%d", #%s)',
-                       ctx:libfunc('string.format'), schema.maxLength, ctx:param(1)))
+      ctx:stmt(sformat('  if length > %d then', schema.maxLength))
+      ctx:stmt(sformat('    return false, %s("string too long, expected at most %d, got %%d", length)',
+                       ctx:libfunc('string.format'), schema.maxLength))
       ctx:stmt(        '  end')
     end
     if schema.pattern then
@@ -1062,6 +1071,9 @@ local _M = {
   -- the ECMA-262 specification but Lua pattern matching library is much more
   -- primitive than that. Users might want to use PCRE or other more powerful
   -- libraries here. The function signature should be: `function(string, patt)`
+  -- @tparam[opt] function custom.str_len function called to get the length of a string. The default
+  -- implementation is `utf8.len` on Lua 5.3+ with a fallback to byte-count if the sequence is invalid UTF-8.
+  -- A custom Lua function is included for older Lua versions. The function signature should be: `function(string)`
   -- @tparam[opt] function custom.external_resolver this will be called to resolve external schemas. It is called with the full
   -- url to fetch (without the fragment part) and must return the
   -- corresponding schema as a Lua table.
@@ -1080,10 +1092,12 @@ local _M = {
     if custom and custom.array_mt ~= nil then
       array_mt = custom.array_mt
     end
+    local str_len = custom and custom.str_len or default_len
     local customlib = {
       null = custom and custom.null or default_null,
       array_mt = array_mt,
-      match_pattern = custom and custom.match_pattern or default_match_pattern
+      match_pattern = custom and custom.match_pattern or default_match_pattern,
+      str_len = str_len,
     }
     local name = custom and custom.name
     return generate_main_validator_ctx(schema, custom):as_func(name, validatorlib, customlib)